+++ jeap-deploymentlog-domain/src/main/java/ch/admin/bit/jeap/deploymentlog/domain/FlowRepository.java
@@ -1,5 +1,6 @@
 package ch.admin.bit.jeap.deploymentlog.domain;
 
+import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -9,6 +10,10 @@ public interface FlowRepository {
     void lockComponent(UUID componentId);
 
     List<Flow> findOpenFlows(UUID componentId, String versionName);
+
+    List<Flow> findOlderOpenFlows(UUID componentId, ZonedDateTime committedBefore, UUID excludedFlowId);
+
+    Optional<Flow> findByDeploymentIdAndLockComponent(UUID deploymentId);
 
     Optional<Flow> findByDeploymentId(UUID deploymentId);