+++ jeap-deploymentlog-domain/src/main/java/ch/admin/bit/jeap/deploymentlog/domain/FlowStageProperties.java
@@ -11,6 +11,7 @@ import org.springframework.stereotype.Component;
 @ConfigurationProperties(prefix = "jeap.deploymentlog.flow", ignoreUnknownFields = false)
 public class FlowStageProperties {
 
+    private boolean enabled = true;
     private String startEnvironment;
     private String defaultFinalDeploymentEnvironment;
 }