+++ jeap-deploymentlog-persistence/src/main/java/ch/admin/bit/jeap/deploymentlog/persistence/JpaComponentRepository.java
@@ -14,4 +14,18 @@ interface JpaComponentRepository extends CrudRepository<Component, UUID> {
 
     @Query(value = "select id from component where id = :id for update", nativeQuery = true)
     Optional<UUID> lockById(@Param("id") UUID id);
+
+    @Query(value = """
+            select component.id
+            from component
+            where component.id = (
+                select component_version.component_id
+                from component_version
+                join flow on flow.component_version_id = component_version.id
+                join deployment on deployment.flow_id = flow.id
+                where deployment.id = :deploymentId
+            )
+            for update
+            """, nativeQuery = true)
+    Optional<UUID> lockByDeploymentId(@Param("deploymentId") UUID deploymentId);
 }